requirements.txt
django
django-jsonfield
django-admin-tools
flake8
pytz
python-dateutil